Jest mock password
